requirements.txt
elevenlabs>=0.2.24
flask>=2.0.0
python-dotenv>=0.19.0